+++ controllers/admin/AdminWorldlineopConfigurationController.php
@@ -136,6 +136,8 @@ class AdminWorldlineopConfigurationController extends ModuleAdminController
 
         $form['webhookMode'] = ($form['webhookMode'] === '1') ? 'automatic' : (($form['webhookMode'] === '0') ? 'manual' : $form['webhookMode']);
 
+        $form = $updater->forceResolve($form);
+
         try {
             $updater->update($form);
         } catch (ExceptionList $e) {