+++ src/Utils/Tools.php
@@ -248,7 +248,8 @@ class Tools
             ->from('orders')
             ->where('id_cart = ' . (int) $idCart);
 
-        $results = \Db::getInstance(_PS_USE_SQL_SLAVE_)->executeS($dbQuery);
+        // Always get order ids from master and make sure that DB query cache is not affecting the query outcome
+        $results = \Db::getInstance()->executeS($dbQuery, true, false);
         if (!$results) {
             return false;
         }
@@ -293,5 +294,23 @@ class Tools
             ->where('id_product = ' . (int) $idProduct);
 
         return \Db::getInstance()->getValue($dbQuery) ?: HostedPaymentRequestBuilder::GIFT_CARD_PRODUCT_TYPE_NONE;
+    }
+
+    /**
+     * @param string|null $env
+     * @return void
+     * @throws \PrestaShopException
+     */
+    public static function removeSymfonyCache($env = null)
+    {
+        if (null === $env) {
+            $env = _PS_ENV_;
+        }
+        $dir = _PS_ROOT_DIR_ . '/var/cache/' . $env .'/';
+        register_shutdown_function(function () use ($dir) {
+            $fs = new Filesystem();
+            $fs->remove($dir);
+            \Hook::exec('actionClearSf2Cache');
+        });
     }
 }