+++ CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.4.12] - 2025-30-09
+## [1.4.13] - 2025-03-10
+### Changed
+- Fix issue showing Mealvouchers on the hosted page
+- Fix saving masked values and API credentials validation
+- Fix database transaction caching issue
+
+## [1.4.12] - 2025-28-08
 ### Added
 - Add configuration to send webhook url in payloads
 
-## [1.4.11] - 2025-30-09
+## [1.4.11] - 2025-22-08
 ### Changed
 - Fix wrong shipping calculation upon clicking back and changing shipping carrier